+++ src/shared/ui/dialog/DialogHeader.tsx
@@ -1,0 +1,4 @@
+export const DialogHeader = ({ className, ...props }) => (
+  <div className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`} {...props} />
+)
+DialogHeader.displayName = "DialogHeader"
